+++ src/main/java/team/tripbeats/backend/repository/MusicRepository.java
@@ -3,5 +3,5 @@ package team.tripbeats.backend.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import team.tripbeats.backend.entity.Music;
 
-public interface MusicRepository extends JpaRepository<Music, Integer> {
+public interface MusicRepository extends JpaRepository<Music, Long> {
 }
