+++ src/main/java/team/tripbeats/backend/repository/TripRepository.java
@@ -5,6 +5,6 @@ import team.tripbeats.backend.entity.Trip;
 
 import java.util.List;
 
-public interface TripRepository extends JpaRepository<Trip, Integer> {
+public interface TripRepository extends JpaRepository<Trip, Long> {
     List<Trip> findByAccountId(Long accountId);
 }
